TR/TRT now has fixed table length of 256
Also reinstated the target length hint on a bunch of instructions.
This is a cosmetic change because the length is only displayed in
the disassembly listing to the right of the instruction format
for information purposes. The length(s) actually assigned to the
instruction target and/or source operands is computed using the
hint() function specified on the instruction format definitions.